fix#13 컴포넌트 타입 위치 변경
