+++ src/components/tabs/tabs.tsx
@@ -1,7 +1,16 @@
 import clsx from 'clsx';
 import { TabsProvider } from './context';
-import { TabsComponentProps } from './type';
 import { findFirstTabId } from './utils';
+import { ReactElement } from 'react';
+import { TabGroupProps } from './subcomponents/tab-group';
+import { ContentGroupProps } from './subcomponents/content-group';
+
+type TabsComponentProps = {
+  children: [ReactElement<TabGroupProps>, ReactElement<ContentGroupProps>];
+  defaultTab?: string;
+  tab?: string;
+  onChange?: (newTab: string) => void;
+};
 
 const TabsComponent = ({ children, defaultTab, ...props }: TabsComponentProps) => {
   const initialTab = defaultTab === undefined ? findFirstTabId(children) : defaultTab;
